clean up test repo
